refactor ♻️: Add Git history and diff analysis, remote management, stash operations, and status tracking
This commit introduces several new modules in the `git_acp` repository: `history.py`, `remote.py`, `runner.py`, `stash.py`, `status.py`, and `tag.py`. These modules provide low-level functionalities for managing git commands, such as executing Git commands, handling signals, and managing stashes. Additionally, a new configuration file is added to specify the number of recent commits to retrieve. The commit messages are also updated to follow a consistent style, ensuring that the repository's commit history remains clear and organized.
