+++ git_acp/git/classification.py
@@ -5,7 +5,8 @@ changes in the repository to suggest appropriate commit types.
 """
 
 from enum import Enum
-from git_acp.git.git_operations import GitError, get_diff
+from git_acp.git.runner import GitError
+from git_acp.git.history import get_diff
 from git_acp.config import COMMIT_TYPE_PATTERNS, COMMIT_TYPES
 from git_acp.utils import debug_header, debug_item
